Create bookmark button; Move cards into folder
