+++ web-app/src/components/cards/Humidity.vue
@@ -2,38 +2,37 @@
   <div class="card-container">
     <ModeButton :value="mode" @input="setMode" />
     <TimeButtons v-model="timeAgo" :zoomed-in="zoomedIn" @reset-zoom="resetZoom()" />
-    <ul class="estimation" v-if="mode === 'current' && measurements.length">
-      <li>
-        <h3>Estimated</h3>
-        <h2>{{ averageElevation }} meters</h2>
-      </li>
-    </ul>
-    <ul class="estimation" v-else-if="mode === 'current'">
-      <li>
-        <h3>Estimated</h3>
-        <h2>N/A</h2>
-      </li>
-    </ul>
+    <CurrentStats v-if="mode === 'current' && measurements.length">
+      <template v-slot:realtime>{{ currentHumidity }}%</template>
+      <template v-slot:average>{{ averageHumidity }}%</template>
+    </CurrentStats>
+    <CurrentStats v-else-if="mode === 'current'">
+      <template v-slot:realtime>N/A</template>
+      <template v-slot:average>N/A</template>
+    </CurrentStats>
     <Graph
       v-else
       ref="graph"
-      chart-type="area"
       :name="sensor.label"
       :measurements="measurements"
-      :options="chartOptions"
       :sensor-type="sensor.type"
       @zoomed-in="zoomedIn = true"
       />
+    <BookmarkButton />
   </div>
 </template>
 
 <script>
-import Graph from './Graph'
-import ModeButton from './ModeButton'
-import TimeButtons from './TimeButtons'
+import BookmarkButton from '../BookmarkButton'
+import CurrentStats from '../CurrentStats'
+import Graph from '../Graph'
+import ModeButton from '../ModeButton'
+import TimeButtons from '../TimeButtons'
 
 export default {
   components: {
+    BookmarkButton,
+    CurrentStats,
     Graph,
     ModeButton,
     TimeButtons
@@ -46,20 +45,21 @@ export default {
   },
   data() {
     return {
-      chartOptions: {
-        stroke: {
-          show: false
-        }
-      },
       mode: 'current',
-      timeAgo: Infinity,
+      timeAgo: 1728e5,
       zoomedIn: false
     }
   },
   computed: {
-    averageElevation() {
+    averageHumidity() {
       const sum = this.measurements.reduce((acc, el) => acc + Number(el.value), 0)
       return Math.round((sum / this.measurements.length) * 10) / 10
+    },
+    currentHumidity() {
+      if (this.measurements.length) {
+        return this.measurements[this.measurements.length - 1].value
+      }
+      return 0
     },
     measurements() {
       if (this.timeAgo === Infinity) {
@@ -84,25 +84,3 @@ export default {
   }
 }
 </script>
-
-<style scoped>
-.card-container {
-  height: 100%;
-}
-
-.estimation {
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-  height: 100%;
-}
-
-.estimation li {
-  align-content: center;
-  display: flex;
-  flex-basis: 100%;
-  flex-direction: column;
-  justify-content: center;
-  text-align: center;
-}
-</style>
